[sapphire] More refactoring
