+++ sapphire/core.py
@@ -9,21 +9,16 @@ Sapphire HTTP server
 import argparse
 import errno
 import logging
-import mimetypes
 import os
 import random
-import re
 import shutil
 import socket
-import sys
 import tempfile
 import threading
 import time
 import traceback
 
 from .sapphire_job import SapphireJob
-from .sapphire_worker import SapphireWorker
-from .server_map import Resource
 from .status_codes import SERVED_ALL, SERVED_NONE, SERVED_TIMEOUT
 
 
@@ -36,15 +31,12 @@ LOG = logging.getLogger("sapphire")
 
 class Sapphire(object):
     ABORT_ON_THREAD_ERROR = False
-    CLOSE_CLIENT_ERROR = None  # used to automatically close client error (4XX code) pages
-    DEFAULT_REQUEST_LIMIT = 0x1000  # 4KB
-    DEFAULT_TX_SIZE = 0x10000  # 64KB
     SHUTDOWN_DELAY = 0.25  # allow extra time before closing socket if needed
-    WORKER_POOL_LIMIT = 10
 
-    _request = re.compile(b"^GET\\s/(?P<request>\\S*)\\sHTTP/1")
-
-    def __init__(self, allow_remote=False, port=None, timeout=60):
+    def __init__(self, allow_remote=False, auto_close=-1, max_workers=10, port=None, timeout=60):
+        assert max_workers > 0
+        self._auto_close = auto_close  # call 'window.close()' on 4xx error pages
+        self._max_workers = max_workers  # limit worker threads
         self._socket = Sapphire._create_listening_socket(allow_remote, port)
         self._timeout = None
         self.timeout = timeout
@@ -54,37 +46,6 @@ class Sapphire(object):
 
     def __exit__(self, *exc):
         self.close()
-
-    @staticmethod
-    def _200_header(c_length, c_type):
-        return "HTTP/1.1 200 OK\r\n" \
-               "Cache-Control: max-age=0, no-cache\r\n" \
-               "Content-Length: %s\r\n" \
-               "Content-Type: %s\r\n" \
-               "Connection: close\r\n\r\n" % (c_length, c_type)
-
-    @staticmethod
-    def _307_redirect(redirct_to):
-        return "HTTP/1.1 307 Temporary Redirect\r\n" \
-               "Location: %s\r\n" \
-               "Connection: close\r\n\r\n" % (redirct_to,)
-
-    @staticmethod
-    def _4xx_page(code, hdr_msg):
-        assert 399 < code < 500
-        if Sapphire.CLOSE_CLIENT_ERROR is not None:
-            assert Sapphire.CLOSE_CLIENT_ERROR >= 0
-            close_timeout = Sapphire.CLOSE_CLIENT_ERROR
-            content = "<script>window.setTimeout(window.close, %d)</script>\n" \
-                      "<body style=\"background-color:#ffffe0\">\n" \
-                      "<h3>%d! - Calling window.close() in %0.1f seconds</h3>\n" \
-                      "</body>\n" % (int(close_timeout * 1000), code, close_timeout)
-        else:
-            content = "<h3>%d!</h3>" % code
-        return "HTTP/1.1 %d %s\r\n" \
-               "Content-Length: %d\r\n" \
-               "Content-Type: text/html\r\n" \
-               "Connection: close\r\n\r\n%s" % (code, hdr_msg, len(content), content)
 
     @staticmethod
     def _create_listening_socket(allow_remote, requested_port):
@@ -121,165 +82,6 @@ class Sapphire(object):
         if self._socket is not None:
             self._socket.close()
 
-    @staticmethod
-    def _handle_request(conn, serv_job):
-        finish_job = False  # call finish() on return
-        try:
-            # receive all the incoming data
-            raw_request = conn.recv(Sapphire.DEFAULT_REQUEST_LIMIT)
-            if not raw_request:
-                LOG.debug("raw_request was empty")
-                serv_job.accepting.set()
-                return
-
-            request = Sapphire._request.match(raw_request)
-            if request is None:
-                serv_job.accepting.set()
-                conn.sendall(Sapphire._4xx_page(400, "Bad Request").encode("ascii"))
-                LOG.debug(
-                    "400 request length %d (%d to go)",
-                    len(raw_request),
-                    serv_job.pending)
-                return
-
-            request = request.group("request").decode("ascii")
-            LOG.debug("check_request(%r)", request)
-            resource = serv_job.check_request(request)
-            if resource is None:
-                LOG.debug("resource is None")  # 404
-            elif resource.type in (Resource.URL_FILE, Resource.URL_INCLUDE):
-                finish_job = serv_job.remove_pending(resource.target)
-            elif resource.type == Resource.URL_REDIRECT:
-                finish_job = serv_job.remove_pending(request)
-
-            if finish_job and serv_job.forever:
-                LOG.debug("serv_job.forever is set, resetting finish_job")
-                finish_job = False
-
-            if not finish_job:
-                serv_job.accepting.set()
-            else:
-                LOG.debug("expecting to finish")
-
-            if resource is None:
-                conn.sendall(Sapphire._4xx_page(404, "Not Found").encode("ascii"))
-                LOG.debug("404 %r (%d to go)", request, serv_job.pending)
-                return
-            if resource.type in (Resource.URL_FILE, Resource.URL_INCLUDE):
-                LOG.debug("target %r", resource.target)
-                if not os.path.isfile(resource.target):
-                    conn.sendall(Sapphire._4xx_page(404, "Not Found").encode("ascii"))
-                    LOG.debug("404 %r (%d to go)", request, serv_job.pending)
-                    return
-                if serv_job.is_forbidden(resource.target):
-                    # NOTE: this does info leak if files exist on disk.
-                    # We could replace 403 with 404 if it turns out we care but this
-                    # is meant to run locally and only be accessible from localhost
-                    conn.sendall(Sapphire._4xx_page(403, "Forbidden").encode("ascii"))
-                    LOG.debug("403 %r (%d to go)", request, serv_job.pending)
-                    return
-            elif resource.type == Resource.URL_REDIRECT:
-                conn.sendall(Sapphire._307_redirect(resource.target).encode("ascii"))
-                LOG.debug(
-                    "307 %r -> %r (%d to go)",
-                    request,
-                    resource.target,
-                    serv_job.pending)
-                return
-            elif resource.type == Resource.URL_DYNAMIC:
-                data = resource.target()
-                if not isinstance(data, bytes):
-                    LOG.debug("dynamic request: %r", request)
-                    raise TypeError("dynamic request callback must return 'bytes'")
-                conn.sendall(Sapphire._200_header(len(data), resource.mime).encode("ascii"))
-                conn.sendall(data)
-                LOG.debug("200 %r (dynamic request)", request)
-                return
-            else:
-                raise RuntimeError("Unknown resource type %r" % resource.type)
-
-            # at this point we know "resource.target" maps to a file on disk
-            # default to "application/octet-stream"
-            c_type = mimetypes.guess_type(resource.target)[0] or "application/octet-stream"
-            # serve the file
-            data_size = os.stat(resource.target).st_size
-            LOG.debug("sending file: %s bytes", format(data_size, ","))
-            with open(resource.target, "rb") as in_fp:
-                conn.sendall(Sapphire._200_header(data_size, c_type).encode("ascii"))
-                offset = 0
-                while offset < data_size:
-                    conn.sendall(in_fp.read(Sapphire.DEFAULT_TX_SIZE))
-                    offset = in_fp.tell()
-            LOG.debug("200 %r (%d to go)", resource.target, serv_job.pending)
-            serv_job.increment_served(resource.target)
-
-        except (socket.timeout, socket.error):
-            exc_type, exc_obj, exc_tb = sys.exc_info()
-            LOG.debug("%s: %r (line %d)", exc_type.__name__, exc_obj, exc_tb.tb_lineno)
-            if not finish_job:
-                serv_job.accepting.set()
-
-        except Exception:  # pylint: disable=broad-except
-            serv_job.exceptions.put(sys.exc_info())
-
-        finally:
-            conn.close()
-            if finish_job:
-                serv_job.finish()
-            serv_job.worker_complete.set()
-
-    @staticmethod
-    def _client_listener(serv_sock, serv_job):
-        worker_pool = list()
-        pool_size = 0
-
-        LOG.debug("starting client_listener")
-        try:
-            while not serv_job.is_complete():
-                if not serv_job.accepting.wait(0.05):
-                    continue
-                try:
-                    worker = SapphireWorker.launch(serv_sock, serv_job)
-                    if worker is not None:
-                        worker_pool.append(worker)
-                        pool_size += 1
-                except threading.ThreadError:
-                    LOG.warning(
-                        "ThreadError! pool size: %d, total active threads: %d",
-                        pool_size,
-                        threading.active_count())
-                    if Sapphire.ABORT_ON_THREAD_ERROR:
-                        raise
-                    # wait for system resources to free up
-                    time.sleep(0.1)
-                # manage worker pool
-                if pool_size >= Sapphire.WORKER_POOL_LIMIT:
-                    LOG.debug("active pool size: %d, waiting for worker to finish...", pool_size)
-                    serv_job.worker_complete.wait()
-                    serv_job.worker_complete.clear()
-                    # remove complete workers
-                    LOG.debug("trimming worker pool")
-                    # sometimes the thread that triggered the event doesn't quite cleanup in time
-                    # so add a retry (10x with a 0.1 second sleep on failure)
-                    for _ in range(10, 0, -1):
-                        worker_pool = list(w for w in worker_pool if not w.done)
-                        pool_size = len(worker_pool)
-                        if pool_size < Sapphire.WORKER_POOL_LIMIT:
-                            break
-                        time.sleep(0.1)
-                    else:
-                        raise RuntimeError("Failed to trim worker pool!")
-                    LOG.debug("trimmed worker pool (size: %d)", pool_size)
-        finally:
-            LOG.debug("shutting down and cleaning up workers")
-            deadline = time.time() + Sapphire.SHUTDOWN_DELAY
-            for worker in worker_pool:
-                # avoid cutting off connections
-                while not worker.done and time.time() < deadline:
-                    LOG.debug("delaying shutdown...")
-                    time.sleep(0.01)
-                worker.close()
-
     @property
     def port(self):
         """
@@ -314,7 +116,12 @@ class Sapphire(object):
         if not os.path.isdir(os.path.abspath(path)):
             raise IOError("%r does not exist" % (path,))
 
-        job = SapphireJob(path, forever=forever, optional_files=optional_files, server_map=server_map)
+        job = SapphireJob(
+            path,
+            auto_close=self._auto_close,
+            forever=forever,
+            optional_files=optional_files,
+            server_map=server_map)
 
         if not job.pending:
             job.finish()
@@ -322,8 +129,9 @@ class Sapphire(object):
 
         # create the client listener thread to handle incoming requests
         listener = threading.Thread(
-            target=self._client_listener,
-            args=(self._socket, job))
+            target=job.client_listener,
+            args=(self._socket, job, self._max_workers),
+            kwargs={"raise_thread_error": self.ABORT_ON_THREAD_ERROR, "shutdown_delay": 0})
 
         # launch listener thread and handle thread errors
         # thread errors can be due to low system resources while fuzzing
